Convert Update objects to ensure type safety on creation
Instead of using generic types in the Update object, we instead check on
instantiation that the type of the object is appropriate for the
UpdateType of the Update. This makes it so if wrong data is sent by an
update, an exception is thrown when the update is created, not when it
is consumed by receiver. This is done to make the tracing of bugs
easier, because Updates are sent across threads, one cannot simply
follow the stack trace to determine where the object was created. By
failing on creation of the object, it is much easier to know where the
faulty code is.

* Add TO_CONTROL_DATA_TYPE and TO_GUI_DATA_TYPE fields to
UpdateType.java
* Add getDataType(boolean isToControl) to UpdateType.java to retrieve
the datatype for the UpdateType and direction
* Convert generic type data to type Object in Update.java
* Add type verification to Update.java constructors