Add try/catch Block to catch Update Creation Errors
In the SwingWorker thread, no errors are thrown by the doInBackground
thread, unless manually caught. So there are no notifications when an
error occurs. This can lead to nasty bugs, because when an error is
thrown, the background process exits, so all updates to the controller
stop getting processed.

By adding a try/catch block, and handling the exception, it is known
when an argument for Update is forgotten.