Relocate wait in doInBackground so more updates are processed
Sometimes multiple updates are sent to doInBackground(), and not all of
them are processed before the thread waits again (e.g. when a new game
starts, both the player name and the signal to start a new game are sent
in quick succession). By moving the wait to the end of the synchronized
block, it appears to process more of the updates before waiting.