Reset number of turns on screen when a new game starts
Previously, an update for the number of turns was not sent when a new
game starts. However, this means that when playing the game for a second
time, the old number of turns will still be there. Now, control sends a
GET_NUM_OF_TURNS update when a new game happens.